[ui, fixup] Fix server render error on unit details page
This commit addresses the following error:
Functions cannot be passed directly to Client Components unless you explicitly expose it by marking it with "use server".

The fix:
Extract legal unit set-primary server-action to separate server-actions file marked with "use server"
